@@ -11,51 +11,45 @@
         "e2e": "ng e2e",
         "doc": "typedoc --options typedoc.json --exclude '**/*.spec.ts' ./src/"
     },
-    "private": true,
+    "private": false,
     "dependencies": {
         "@angular-redux/store": "6.5.7",
-        "@angular/animations": "4.4.7",
-        "@angular/cdk": "5.2.5",
-        "@angular/common": "4.4.7",
-        "@angular/compiler": "4.4.7",
-        "@angular/core": "4.4.7",
-        "@angular/forms": "4.4.7",
-        "@angular/http": "4.4.7",
-        "@angular/platform-browser": "4.4.7",
-        "@angular/platform-browser-dynamic": "4.4.7",
-        "@angular/router": "4.4.7",
+        "@angular/animations": "6.0.6",
+        "@angular/cdk": "6.2.1",
+        "@angular/common": "6.0.6",
+        "@angular/core": "6.0.6",
+        "@angular/forms": "6.0.6",
+        "@angular/http": "6.0.6",
+        "@angular/platform-browser": "6.0.6",
+        "@angular/platform-browser-dynamic": "6.0.6",
+        "@angular/router": "6.0.6",
         "angular2-hotkeys": "2.0.4",
-        "bootstrap": "4.0.0-alpha.6",
-        "core-js": "2.5.4",
+        "bootstrap": "4.1.1",
+        "core-js": "2.5.6",
         "css-element-queries": "0.4.0",
         "element-resize-detector": "1.1.12",
         "elkjs": "0.0.1",
         "font-awesome": "4.7.0",
-        "jsplumb": "2.6.9",
-        "ng2-toastr": "4.1.2",
+        "jsplumb": "2.7.9",
         "ngx-bootstrap": "2.0.4",
+        "ngx-toastr": "8.5.1",
         "redux": "3.7.2",
-        "rxjs": "5.5.8",
-        "zone.js": "0.8.25"
+        "rxjs": "6.1.0",
+        "rxjs-compat": "6.2.1",
+        "zone.js": "0.8.26"
     },
     "devDependencies": {
-        "@angular/cli": "1.6.6",
-        "@angular/compiler-cli": "4.4.7",
-        "@types/jasmine": "2.8.6",
-        "@types/node": "9.6.2",
-        "codelyzer": "3.1.2",
-        "jasmine-core": "2.7.0",
-        "jasmine-spec-reporter": "4.2.1",
-        "karma": "1.7.0",
-        "karma-chrome-launcher": "2.2.0",
-        "karma-cli": "1.0.1",
-        "karma-coverage-istanbul-reporter": "1.3.0",
-        "karma-jasmine": "1.1.0",
-        "karma-jasmine-html-reporter": "0.2.2",
-        "protractor": "5.1.2",
-        "ts-node": "3.3.0",
-        "tslint": "5.6.0",
-        "typedoc": "0.11.1",
-        "typescript": "2.4.2"
+        "@angular/cli": "6.0.8",
+        "@angular/compiler": "6.0.6",
+        "@angular/compiler-cli": "6.0.6",
+        "@angular-devkit/build-angular": "0.6.8",
+        "@angular-devkit/core": "0.6.8",
+        "typescript": "2.7.2",
+        "@types/node": "10.3.2",
+        "codelyzer": "4.2.1",
+        "protractor": "5.3.0",
+        "ts-node": "5.0.1",
+        "tslib": "1.9.2",
+        "tslint": "5.10.0"
     }
 }
