@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017-2018 Contributors to the Eclipse Foundation
+ * Copyright (c) 2017 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -11,12 +11,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
-import {Component, OnInit, ViewChild, ViewContainerRef} from '@angular/core';
-import {WineryNotificationService} from './wineryNotificationModule/wineryNotification.service';
-import {WineryGitLogComponent} from './wineryGitLog/wineryGitLog.component';
-import {ExistService} from './wineryUtils/existService';
-import {backendBaseURL} from './configuration';
-import {BackendAvailabilityStates} from './model/enums';
+import { Component, OnInit } from '@angular/core';
+import { WineryNotificationService } from './wineryNotificationModule/wineryNotification.service';
+import { ExistService } from './wineryUtils/existService';
+import { backendBaseURL } from './configuration';
+import { BackendAvailabilityStates } from './model/enums';
 
 @Component({
     selector: 'winery-repository',
@@ -30,12 +29,13 @@ import {BackendAvailabilityStates} from './model/enums';
  * This component represents the root component for the Winery Repository.
  */
 export class WineryRepositoryComponent implements OnInit {
+
     // region variables
     name = 'Winery Repository';
     backendState = BackendAvailabilityStates.Undefined;
     backendAvailabilityStates = BackendAvailabilityStates;
     loading = true;
-    @ViewChild('gitLog') gitLog: WineryGitLogComponent;
+    // @ViewChild('gitLog') gitLog: WineryGitLogComponent;
 
     // endregion
     options = {
@@ -61,9 +61,9 @@ export class WineryRepositoryComponent implements OnInit {
         );
     }
 
-    onClick() {
-        // this.gitLog.hide();
-    }
+    // onClick() {
+    // this.gitLog.hide();
+    // }
 
     refresh() {
         window.location.reload();
