@@ -42,7 +42,8 @@ class CopybaraGeneratorTest extends TestWithGitBackedRepository {
             "        push = \"master\",\n" +
             "    ),\n" +
             "    authoring = authoring.pass_thru(\"OpenTOSCA Bot <opentosca@iaas.uni-stuttgart.de>\"),\n" +
-            "    origin_files = glob([\"README.md\", \"LICENSE\", \"artifacttemplates/http%3A%2F%2Fplain.winery.opentosca.org%2Fartifacttemplates/ArtifactTemplateWithFilesAndSources-ArtifactTypeWithoutProperties/**\",\n" +
+            "    origin_files = glob([\"README.md\", \"LICENSE\", \".circleci/config.yml\",\n" +
+            "        \"artifacttemplates/http%3A%2F%2Fplain.winery.opentosca.org%2Fartifacttemplates/ArtifactTemplateWithFilesAndSources-ArtifactTypeWithoutProperties/**\",\n" +
             "        \"artifacttemplates/http%3A%2F%2Fplain.winery.opentosca.org%2Fartifacttemplates/ArtifactTemplateWithoutAnyFiles-ArtifactTypeWithoutProperties/**\"]),\n" +
             "    destination_files = glob([\"**\"], exclude = [\"README_INTERNAL.md\"]),\n" +
             ")", config);
@@ -54,7 +55,8 @@ class CopybaraGeneratorTest extends TestWithGitBackedRepository {
         this.setRevisionTo("4b218e6");
         CopybaraGenerator copybaraGenerator = new CopybaraGenerator();
         String config = copybaraGenerator.generateOriginFilesConfig();
-        assertEquals("origin_files = glob([\"README.md\", \"LICENSE\", \"artifacttemplates/http%3A%2F%2Fplain.winery.opentosca.org%2Fartifacttemplates/ArtifactTemplateWithFilesAndSources-ArtifactTypeWithoutProperties/**\",\n" +
+        assertEquals("origin_files = glob([\"README.md\", \"LICENSE\", \".circleci/config.yml\",\n" +
+            "        \"artifacttemplates/http%3A%2F%2Fplain.winery.opentosca.org%2Fartifacttemplates/ArtifactTemplateWithFilesAndSources-ArtifactTypeWithoutProperties/**\",\n" +
             "        \"artifacttemplates/http%3A%2F%2Fplain.winery.opentosca.org%2Fartifacttemplates/ArtifactTemplateWithoutAnyFiles-ArtifactTypeWithoutProperties/**\"]),", config);
     }
 }
