@@ -12,10 +12,10 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  ********************************************************************************/
 import {
-    AfterViewInit, Component, DoCheck, ElementRef, HostListener, Input, KeyValueDiffers, NgZone, OnDestroy, OnInit,
-    QueryList, Renderer2, ViewChild, ViewChildren
+    AfterViewInit, Component, DoCheck, ElementRef, HostListener, Input, KeyValueDiffers, NgZone, OnChanges, OnDestroy, OnInit, QueryList,
+    Renderer2, SimpleChanges, ViewChild, ViewChildren
 } from '@angular/core';
-import { JsPlumbService } from '../services/jsPlumbService';
+import { JsPlumbService } from '../services/jsPlumb.service';
 import { EntityType, TNodeTemplate, TRelationshipTemplate } from '../models/ttopology-template';
 import { LayoutDirective } from '../layout/layout.directive';
 import { WineryActions } from '../redux/actions/winery.actions';
@@ -26,13 +26,13 @@ import { TopologyRendererActions } from '../redux/actions/topologyRenderer.actio
 import { NodeComponent } from '../node/node.component';
 import { Hotkey, HotkeysService } from 'angular2-hotkeys';
 import { ModalDirective } from 'ngx-bootstrap';
-import { GridTemplate } from 'app/models/gridTemplate';
-import { Subscription } from 'rxjs/Subscription';
+import { GridTemplate } from '../models/gridTemplate';
+import { Subscription } from 'rxjs';
 import { CapabilitiesModalData } from '../models/capabilitiesModalData';
 import { RequirementsModalData } from '../models/requirementsModalData';
 import { NodeIdAndFocusModel } from '../models/nodeIdAndFocusModel';
 import { ToggleModalDataModel } from '../models/toggleModalDataModel';
-import { WineryAlertService } from '../winery-alert/winery-alert.service';
+import { ToastrService } from 'ngx-toastr';
 import { BackendService } from '../services/backend.service';
 import { hostURL } from '../models/configuration';
 import { CapabilityModel } from '../models/capabilityModel';
@@ -57,7 +57,7 @@ import { DragSource } from '../models/DragSource';
     templateUrl: './canvas.component.html',
     styleUrls: ['./canvas.component.css']
 })
-export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoCheck {
+export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewInit, DoCheck {
 
     @ViewChildren(NodeComponent) nodeComponentChildren: QueryList<NodeComponent>;
     @ViewChildren('KVTextareas') KVTextareas: QueryList<any>;
@@ -67,9 +67,11 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
     @ViewChild('capabilitiesModal') capabilitiesModal: ModalDirective;
     @ViewChild('requirementsModal') requirementsModal: ModalDirective;
     @ViewChild('importTopologyModal') importTopologyModal: ModalDirective;
+    @Input() readonly: boolean;
     @Input() entityTypes: EntityTypesModel;
     @Input() relationshipTypes: Array<EntityType> = [];
     @Input() diffMode = false;
+    @Input() sidebarDeleteButtonClickEvent: any;
 
     readonly draggingThreshold = 300;
     readonly newNodePositionOffsetX = 108;
@@ -82,6 +84,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
     // current data emitted from a node
     currentModalData: any;
     dragSourceActive = false;
+    event;
     selectedRelationshipType: EntityType;
     nodeChildrenIdArray: Array<string>;
     nodeChildrenArray: Array<NodeComponent>;
@@ -144,7 +147,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
                 private zone: NgZone,
                 private hotkeysService: HotkeysService,
                 private renderer: Renderer2,
-                private alert: WineryAlertService,
+                private alert: ToastrService,
                 private differs: KeyValueDiffers,
                 private backendService: BackendService,
                 private importTopologyService: ImportTopologyService,
@@ -166,14 +169,25 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
         this.gridTemplate = new GridTemplate(100, false, false, 30);
         this.subscriptions.push(this.ngRedux.select(state => state.wineryState.currentPaletteOpenedState)
             .subscribe(currentPaletteOpened => this.setPaletteState(currentPaletteOpened)));
-        this.hotkeysService.add(new Hotkey('ctrl+a', (event: KeyboardEvent): boolean => {
+        this.hotkeysService.add(new Hotkey('mod+a', (event: KeyboardEvent): boolean => {
             event.stopPropagation();
             this.allNodeTemplates.forEach(node => this.enhanceDragSelection(node.id));
             return false; // Prevent bubbling
-        }));
+        }, undefined, 'Select all Node Templates'));
+        this.hotkeysService.add(new Hotkey('del', (event: KeyboardEvent): boolean => {
+            this.handleDeleteKeyEvent();
+            return false;
+        }, undefined, 'Delete an element.'));
         this.capabilities = new CapabilitiesModalData();
         this.requirements = new RequirementsModalData();
         this.importTopologyData = new ImportTopologyModalData();
+    }
+
+    ngOnChanges(changes: SimpleChanges) {
+        if (changes) {
+            const buttonClickEvent = changes.sidebarDeleteButtonClickEvent;
+            this.handleDeleteKeyEvent();
+        }
     }
 
     /**
@@ -257,48 +271,6 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
             this.gridTemplate.marginLeft = 300;
         } else {
             this.gridTemplate.marginLeft = 30;
-        }
-    }
-
-    /**
-     * Gets all ID's of the topology template and saves them in an array
-     */
-    private getAllIds(): void {
-        this.allIds = [];
-        // get all Id's of the node templates
-        this.allNodeTemplates.forEach(node => {
-            if (this.allIds.length > 0) {
-                this.setId(node.id);
-                if (node.requirements) {
-                    if (node.requirements.requirement) {
-                        node.requirements.requirement.forEach(req => {
-                            this.setId(req.id);
-                        });
-                    }
-                }
-                if (node.capabilities) {
-                    if (node.capabilities.capability) {
-                        node.capabilities.capability.forEach(cap => {
-                            this.setId(cap.id);
-                        });
-                    }
-                }
-            } else {
-                this.allIds.push(node.id);
-            }
-        });
-        // get all relationship Id's
-        this.allRelationshipTemplates.forEach(rel => {
-            this.setId(rel.id);
-        });
-    }
-
-    /**
-     * Checks if the id is already in the array, if not the id is added
-     */
-    private setId(idOfElement: string): void {
-        if (!this.allIds.find(id => id === idOfElement)) {
-            this.allIds.push(idOfElement);
         }
     }
 
@@ -665,7 +637,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
             };
         }
         // case when there are no capabilities on the node template
-        if (isNullOrUndefined(this.capabilities.capabilities)) {
+        if (!this.capabilities.capabilities || !this.capabilities.capabilities.capability) {
             const capabilityArray: Array<CapabilityModel> = [];
             this.capabilities.capabilities = {
                 capability: capabilityArray
@@ -771,7 +743,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
             };
         }
         // case when there are no requirements on the node template
-        if (isNullOrUndefined(this.requirements.requirements)) {
+        if (!this.requirements.requirements || !this.requirements.requirements.requirement) {
             const requirementsArray: Array<RequirementModel> = [];
             this.requirements.requirements = {
                 requirement: requirementsArray
@@ -1012,6 +984,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
                         for (const serviceTemplate of allServiceTemplates) {
                             this.importTopologyData.allTopologyTemplates.push(serviceTemplate);
                         }
+                        console.log(this.importTopologyData.allTopologyTemplates);
                     });
                 }
                 this.ngRedux.dispatch(this.topologyRendererActions.importTopology());
@@ -1093,15 +1066,15 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
     setNewCoordinates(nodeTemplate: any): void {
         let nodeIndex;
         this.allNodeTemplates.some((node, index) => {
-            if (node.id === nodeTemplate.firstChild.nextElementSibling.id) {
+            if (node.id === nodeTemplate.firstChild.id) {
                 nodeIndex = index;
                 return true;
             }
         });
         const nodeCoordinates = {
-            id: nodeTemplate.firstChild.nextElementSibling.id,
-            x: nodeTemplate.firstChild.nextElementSibling.offsetLeft,
-            y: nodeTemplate.firstChild.nextElementSibling.offsetTop
+            id: nodeTemplate.firstChild.id,
+            x: nodeTemplate.firstChild.offsetLeft,
+            y: nodeTemplate.firstChild.offsetTop
         };
         this.allNodeTemplates[nodeIndex].x = nodeCoordinates.x;
         this.allNodeTemplates[nodeIndex].y = nodeCoordinates.y;
@@ -1114,7 +1087,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
     updateSelectedNodes(): void {
         if (this.selectedNodes.length > 0 && this.child) {
             for (const nodeTemplate of this.child.nativeElement.children) {
-                if (this.selectedNodes.some(node => node.id === nodeTemplate.firstChild.nextElementSibling.id)) {
+                if (this.selectedNodes.some(node => node.id === nodeTemplate.firstChild.id)) {
                     this.setNewCoordinates(nodeTemplate);
                 }
             }
@@ -1139,6 +1112,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
 
             const border = isNullOrUndefined(newRelationship.state)
                 ? '#fafafa' : VersionUtils.getElementColorByDiffState(newRelationship.state);
+            const me = this;
             const conn = this.newJsPlumbInstance.connect({
                 source: newRelationship.sourceElement.ref,
                 target: newRelationship.targetElement.ref,
@@ -1146,6 +1120,11 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
                     ['Label', {
                         label: labelString,
                         id: 'label',
+                        events: {
+                            click: function (labelOverlay, originalEvent) {
+                                setTimeout(() => me.onClickJsPlumbConnection(conn, newRelationship), 1);
+                            }
+                        },
                         labelStyle: {
                             font: '11px Roboto, sans-serif',
                             color: '#212121',
@@ -1257,8 +1236,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
      * Handler for the DEL-Key - removes a node and resets everything associated with that deleted node
      * @param event Keyboard event.
      */
-    @HostListener('document:keydown.delete', ['$event'])
-    handleDeleteKeyEvent(event: KeyboardEvent) {
+    handleDeleteKeyEvent(event?: KeyboardEvent) {
         this.unbindConnection();
         // if name, min or max instances has changed, do not delete the node.
         if (this.selectedNodes.length > 0) {
@@ -1292,7 +1270,6 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
                 }
             }
         }
-
     }
 
     /**
@@ -1372,7 +1349,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
             const bElem = node.firstChild;
             const result = this.isObjectInSelection(aElem, bElem);
             if (result) {
-                this.enhanceDragSelection(node.firstChild.nextElementSibling.id);
+                this.enhanceDragSelection(node.firstChild.id);
             }
         }
         this.unbindAll();
@@ -1430,7 +1407,7 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
 
     /**
      * Checks if array 'Nodes' contains 'id'.
-     * @param Nodes
+     * @param nodes
      * @param id
      * @returns Boolean True if 'Nodes' contains 'id'.
      */
@@ -1516,21 +1493,8 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
                 }
             }
         });
-        this.differ = this.differs.find([]).create(null);
+        this.differ = this.differs.find([]).create();
     }
-
-    /*
-    isFieldValid(field: string) {
-        return !this.form.get(field).valid && this.form.get(field).touched;
-    }
-
-    displayFieldCss(field: string) {
-        return {
-            'has-error': this.isFieldValid(field),
-            'has-feedback': this.isFieldValid(field)
-        };
-    }
-*/
 
     /**
      * Angular lifecycle event.
@@ -1552,6 +1516,19 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
         this.selectedRelationshipType = currentType;
     }
 
+    /*
+    isFieldValid(field: string) {
+        return !this.form.get(field).valid && this.form.get(field).touched;
+    }
+
+    displayFieldCss(field: string) {
+        return {
+            'has-error': this.isFieldValid(field),
+            'has-feedback': this.isFieldValid(field)
+        };
+    }
+*/
+
     /**
      * Removes an element from JSPlumb.
      * @param id
@@ -1566,9 +1543,11 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
      * @param nodeId
      */
     activateNewNode(nodeId: string): void {
-        this.newJsPlumbInstance.draggable(nodeId);
-        if (this.paletteOpened) {
-            this.bindNewNode();
+        if (nodeId) {
+            this.newJsPlumbInstance.draggable(nodeId);
+            if (this.paletteOpened) {
+                this.bindNewNode();
+            }
         }
     }
 
@@ -1643,6 +1622,48 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
      */
     ngOnDestroy() {
         this.subscriptions.forEach(subscription => subscription.unsubscribe());
+    }
+
+    /**
+     * Gets all ID's of the topology template and saves them in an array
+     */
+    private getAllIds(): void {
+        this.allIds = [];
+        // get all Id's of the node templates
+        this.allNodeTemplates.forEach(node => {
+            if (this.allIds.length > 0) {
+                this.setId(node.id);
+                if (node.requirements) {
+                    if (node.requirements.requirement) {
+                        node.requirements.requirement.forEach(req => {
+                            this.setId(req.id);
+                        });
+                    }
+                }
+                if (node.capabilities) {
+                    if (node.capabilities.capability) {
+                        node.capabilities.capability.forEach(cap => {
+                            this.setId(cap.id);
+                        });
+                    }
+                }
+            } else {
+                this.allIds.push(node.id);
+            }
+        });
+        // get all relationship Id's
+        this.allRelationshipTemplates.forEach(rel => {
+            this.setId(rel.id);
+        });
+    }
+
+    /**
+     * Checks if the id is already in the array, if not the id is added
+     */
+    private setId(idOfElement: string): void {
+        if (!this.allIds.find(id => id === idOfElement)) {
+            this.allIds.push(idOfElement);
+        }
     }
 
     /**
@@ -1741,25 +1762,32 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
     private handleRelSideBar(conn: any, newRelationship: TRelationshipTemplate): void {
         conn.id = newRelationship.id;
         conn.setType(newRelationship.type);
-        const me = this;
         conn.bind('click', rel => {
-            this.clearSelectedNodes();
-            this.newJsPlumbInstance.select().removeType('marked');
-            const currentRel = me.allRelationshipTemplates.find(con => con.id === rel.id);
-            if (currentRel) {
-                me.ngRedux.dispatch(this.actions.openSidebar({
-                    sidebarContents: {
-                        sidebarVisible: true,
-                        nodeClicked: false,
-                        id: currentRel.id,
-                        nameTextFieldValue: currentRel.name,
-                        type: currentRel.type
-                    }
-                }));
-                conn.addType('marked');
-            }
+            this.onClickJsPlumbConnection(conn, rel);
         });
+
         this.revalidateContainer();
+    }
+
+    /**
+     * jsPlumb relationship/label click actions
+     */
+    onClickJsPlumbConnection(conn: any, rel: any) {
+        this.clearSelectedNodes();
+        this.newJsPlumbInstance.select().removeType('marked');
+        const currentRel = this.allRelationshipTemplates.find(con => con.id === rel.id);
+        if (currentRel) {
+            this.ngRedux.dispatch(this.actions.openSidebar({
+                sidebarContents: {
+                    sidebarVisible: true,
+                    nodeClicked: false,
+                    id: currentRel.id,
+                    nameTextFieldValue: currentRel.name,
+                    type: currentRel.type
+                }
+            }));
+            conn.addType('marked');
+        }
     }
 
     /**
@@ -1777,12 +1805,12 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
     private isObjectInSelection(selectionArea, object): boolean {
         const selectionRect = selectionArea.getBoundingClientRect();
         return (
-            ((selectionRect.top + selectionRect.height) > (object.nextElementSibling.offsetTop +
-                object.nextElementSibling.offsetHeight - this.scrollOffset)) &&
-            (selectionRect.top < (object.nextElementSibling.offsetTop - this.scrollOffset)) &&
-            ((selectionRect.left + selectionArea.getBoundingClientRect().width) > (object.nextElementSibling.offsetLeft +
-                object.nextElementSibling.offsetWidth)) &&
-            (selectionRect.left < (object.nextElementSibling.offsetLeft))
+            ((selectionRect.top + selectionRect.height) > (object.offsetTop +
+                object.offsetHeight - this.scrollOffset)) &&
+            (selectionRect.top < (object.offsetTop - this.scrollOffset)) &&
+            ((selectionRect.left + selectionArea.getBoundingClientRect().width) > (object.offsetLeft +
+                object.offsetWidth)) &&
+            (selectionRect.left < (object.offsetLeft))
         );
     }
 
@@ -1935,13 +1963,15 @@ export class CanvasComponent implements OnInit, OnDestroy, AfterViewInit, DoChec
      * Handles the new node by binding to mouse move and mouse up actions
      */
     private bindNewNode(): void {
-        setTimeout(() => this.handleNodePressActions(this.newNode.id), 1);
-        this.zone.run(() => {
-            this.unbindMouseActions.push(this.renderer.listen(this.eref.nativeElement, 'mousemove',
-                (event) => this.moveNewNode(event)));
-            this.unbindMouseActions.push(this.renderer.listen(this.eref.nativeElement, 'mouseup',
-                ($event) => this.positionNewNode()));
-        });
+        if (this.newNode) {
+            setTimeout(() => this.handleNodePressActions(this.newNode.id), 1);
+            this.zone.run(() => {
+                this.unbindMouseActions.push(this.renderer.listen(this.eref.nativeElement, 'mousemove',
+                    (event) => this.moveNewNode(event)));
+                this.unbindMouseActions.push(this.renderer.listen(this.eref.nativeElement, 'mouseup',
+                    ($event) => this.positionNewNode()));
+            });
+        }
     }
 
     /**
