@@ -12,7 +12,17 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
 
-import { AfterViewInit, Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
+import {
+    AfterViewInit,
+    Component,
+    EventEmitter,
+    Input,
+    OnChanges,
+    OnInit,
+    Output,
+    SimpleChanges,
+    ViewChild
+} from '@angular/core';
 import { ModalDirective } from 'ngx-bootstrap';
 import { TPolicy } from '../../models/policiesModalData';
 import { TDeploymentArtifact } from '../../models/artifactsModalData';
@@ -25,7 +35,7 @@ import { isNullOrUndefined } from 'util';
 import { backendBaseURL, hostURL } from '../../models/configuration';
 import { WineryActions } from '../../redux/actions/winery.actions';
 import { ExistsService } from '../../services/exists.service';
-import { WineryAlertService } from '../../winery-alert/winery-alert.service';
+import { ToastrService } from 'ngx-toastr';
 import { DeploymentArtifactOrPolicyModalData, ModalVariant, ModalVariantAndState } from './modal-model';
 import { EntitiesModalService, OpenModalEvent } from './entities-modal.service';
 import { QName } from '../../models/qname';
@@ -68,7 +78,7 @@ export class EntitiesModalComponent implements OnInit, AfterViewInit, OnChanges 
                 private actions: WineryActions,
                 private existsService: ExistsService,
                 private entitiesModalService: EntitiesModalService,
-                private alert: WineryAlertService) {
+                private alert: ToastrService) {
     }
 
     ngOnInit() {
