@@ -14,7 +14,7 @@
 
 // when running in development mode on port 4201, use default port 8080
 // otherwise, assume that backend runs on the some port
-export const hostURL = location.protocol + '//' + location.hostname + ':' + (location.port === '4201' ? '8080' : location.port);
+export const hostURL = location.protocol + '//' + location.hostname + ':' + (location.port === '4201' || location.port === '4202' ? '8080' : location.port);
 export const backendBaseURL = hostURL + '/winery';
 // when running in development mode, use the workflow modelers development port
 // it also is not running on /winery-workflowmodeler in dev-mode.
