@@ -17,7 +17,7 @@ import { EntityTypesModel } from '../../models/entityTypesModel';
 import { TNodeTemplate } from '../../models/ttopology-template';
 import { NgRedux } from '@angular-redux/store';
 import { IWineryState } from '../../redux/store/winery.store';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 import { CapabilityModel } from '../../models/capabilityModel';
 
 @Component({
@@ -31,6 +31,7 @@ import { CapabilityModel } from '../../models/capabilityModel';
 export class CapabilitiesComponent implements OnInit, OnChanges, OnDestroy {
 
     @Output() toggleModalHandler: EventEmitter<any>;
+    @Input() readonly: boolean;
     @Input() currentNodeData: any;
 
     capabilities: any[] = [];
