@@ -67,6 +67,7 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
     removeZIndex: any;
     propertyDefinitionType: string;
 
+    @Input() readonly: boolean;
     @Input() entityTypes: EntityTypesModel;
     @Input() dragSource: string;
     @Input() navbarButtonsState: ButtonsStateModel;
@@ -93,6 +94,7 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
     hostURL = hostURL;
     flashTimer = 300;
     parentEl: string;
+    popoverHtml = `<div class="">Open NodeType in a separate tab</div>`;
     // differ object for detecting changes made to the nodeTemplate object for DoCheck
     differ: any;
 
@@ -132,8 +134,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
     /**
      * This function determines which kind of properties the nodeType embodies.
      * We have 3 possibilities: none, XML element, or Key value pairs.
-     * @param {string} type
-     * @param groupedNodeTypes
      */
     findOutPropertyDefinitionTypeForProperties(type: string, groupedNodeTypes: Array<GroupedNodeTypeModel>): void {
         let propertyDefinitionTypeAssigned: boolean;
@@ -168,7 +168,7 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
      * Angular lifecycle event.
      */
     ngOnInit() {
-        this.differ = this.differs.find([]).create(null);
+        this.differ = this.differs.find([]).create();
         this.nodeClass = this.nodeTemplate.visuals.pattern ? 'pattern' : 'nodeTemplate';
     }
 
@@ -203,7 +203,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Stops the event propagation to the canvas etc. and repaints.
-     * @param $event
      */
     repaint($event) {
         $event.stopPropagation();
@@ -212,7 +211,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Sets the current type of a node.
-     * @param $event
      */
     passCurrentType($event): void {
         $event.stopPropagation();
@@ -233,7 +231,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Handler for mousedown events, toggles visibility of node attributes
-     * @param $event
      */
     mouseDownHandler($event): void {
         this.unmarkConnections.emit();
@@ -250,8 +247,8 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
             this.parentEl = $event.target.parentElement.className;
         }
         if (this.parentEl !== 'accordion-toggle' && this.parentEl !== 'ng-tns-c6-2' && this.parentEl) {
-            const offsetLeft = this.elRef.nativeElement.firstChild.nextElementSibling.offsetLeft;
-            const offsetTop = this.elRef.nativeElement.firstChild.nextElementSibling.offsetTop;
+            const offsetLeft = this.elRef.nativeElement.firstChild.offsetLeft;
+            const offsetTop = this.elRef.nativeElement.firstChild.offsetTop;
             this.previousPosition = {
                 x: offsetLeft,
                 y: offsetTop
@@ -264,11 +261,10 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * If a node is moved, this saves the current position of the node into the store.
-     * @param $event
      */
     mouseMove($event): void {
-        const offsetLeft = this.elRef.nativeElement.firstChild.nextElementSibling.offsetLeft;
-        const offsetTop = this.elRef.nativeElement.firstChild.nextElementSibling.offsetTop;
+        const offsetLeft = this.elRef.nativeElement.firstChild.offsetLeft;
+        const offsetTop = this.elRef.nativeElement.firstChild.offsetTop;
         this.currentPosition = {
             x: offsetLeft,
             y: offsetTop
@@ -277,7 +273,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Checks if it was a click or a drag operation on the node.
-     * @param $event
      */
     mouseUpHandler($event): void {
         // mouseup
@@ -313,7 +308,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * If it was a click operation, close the connector endpoints for relations
-     * @param $event
      */
     closeConnectorEndpoints($event): void {
         $event.stopPropagation();
@@ -325,7 +319,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Creates a dragoperation for nodes
-     * @param $event
      */
     makeSource($event): void {
         const dragSourceInfo = {
@@ -337,7 +330,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Only display the sidebar if the click is no longpress (drag)
-     * @param $event
      */
     openSidebar($event): void {
         $event.stopPropagation();
@@ -372,11 +364,11 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Navigates to the corresponding node type in the management UI
-     * @param $event
+     *  $event
      */
     linkType($event: any): void {
         const qName = new QName(this.nodeTemplate.type);
-        const typeURL = this.backendService.configuration.uiURL + '#' + urlElement.NodeTypeURL +
+        const typeURL = this.backendService.configuration.uiURL + urlElement.NodeTypeURL +
             encodeURIComponent(encodeURIComponent(qName.nameSpace)) + '/' + qName.localName
             + urlElement.ReadMe;
         window.open(typeURL, '_blank');
@@ -384,7 +376,6 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Displays a box of the whole text if the text doesn't fit in the original element
-     * @param cell
      */
     isEllipsisActive(cell) {
         return (cell.offsetWidth < cell.scrollWidth);
@@ -402,7 +393,7 @@ export class NodeComponent implements OnInit, AfterViewInit, OnDestroy, DoCheck 
 
     /**
      * Checks if it was a click or a drag operation on the node.
-     * @param $event
+     *  $event
      */
     private testTimeDifference($event): void {
         if ((this.endTime - this.startTime) < 200) {
