@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BackendService } from './backend.service';
-import { WineryAlertService } from '../winery-alert/winery-alert.service';
+import { ToastrService } from 'ngx-toastr';
 import { NgRedux } from '@angular-redux/store';
 import { IWineryState } from '../redux/store/winery.store';
 import { TopologyRendererActions } from '../redux/actions/topologyRenderer.actions';
@@ -9,7 +9,7 @@ import { ErrorHandlerService } from './error-handler.service';
 @Injectable()
 export class SplitMatchTopologyService {
 
-    constructor(private alert: WineryAlertService) {
+    constructor(private alert: ToastrService) {
     }
 
     /**
